+++ Mods/ScribeSkillbooks_ef0ad26a-20e5-4935-9e7e-baef3dff1664/Story/RawFiles/Lua/BootstrapClient.lua
@@ -3,5 +3,6 @@
 --  =======
 
 Ext.Require("Auxiliary.lua")
+Ext.Require("Client/TreasureTables.lua")
 Ext.Require("Client/LegacyCompatibilityMode.lua")
 if MODINFO.ModSettings["RecipeGeneration"] then Ext.Require("Client/RecipeGenerator.lua") end